Profile files moved to other folder
