+++ Source/NETworkManager/MainWindow.xaml.cs
@@ -24,6 +24,7 @@ using NETworkManager.ViewModels;
 using NETworkManager.Models.EventSystem;
 using ContextMenu = System.Windows.Controls.ContextMenu;
 using DnsClient;
+using NETworkManager.Models.Profile;
 
 namespace NETworkManager
 {
