+++ Source/NETworkManager/Models/EventSystem/EventSystemRedirectProfileApplicationArgs.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using NETworkManager.Models.Profile;
 using NETworkManager.Models.Settings;
 
 namespace NETworkManager.Models.EventSystem
