+++ Source/NETworkManager/Models/Network/WakeOnLan.cs
@@ -1,4 +1,5 @@
-﻿using NETworkManager.Models.Settings;
+﻿using NETworkManager.Models.Profile;
+using NETworkManager.Models.Settings;
 using NETworkManager.Utilities;
 using System;
 using System.Net;
