+++ Source/NETworkManager/Models/PuTTY/PuTTYSessionInfo.cs
@@ -1,4 +1,5 @@
-﻿using NETworkManager.Models.Settings;
+﻿using NETworkManager.Models.Profile;
+using NETworkManager.Models.Settings;
 using static NETworkManager.Models.PuTTY.PuTTY;
 
 namespace NETworkManager.Models.PuTTY
