+++ Source/NETworkManager/Resources/Localization/Strings.resx
@@ -2340,8 +2340,8 @@ $$hostname$$	--&gt;	Hostname</value>
   <data name="IPAddressToDetectLocalIPAddressBasedOnRouting" xml:space="preserve">
     <value>IP address to detect local ip address based on routing</value>
   </data>
-  <data name="ShowWindowOnNetworkChange" xml:space="preserve">
-    <value>Show window on network change</value>
+  <data name="ShowStatusWindowOnNetworkChange" xml:space="preserve">
+    <value>Show status window on network change</value>
   </data>
   <data name="TimeInSecondsHowLongTheWindowShouldBeDisplayed" xml:space="preserve">
     <value>Time in seconds how long the window should be displayed</value>
