+++ Source/NETworkManager/ViewModels/PortScannerHostViewModel.cs
@@ -11,6 +11,7 @@ using System.Windows.Data;
 using MahApps.Metro.Controls.Dialogs;
 using System.Windows;
 using System.Linq;
+using NETworkManager.Models.Profile;
 
 namespace NETworkManager.ViewModels
 {
