+++ Source/NETworkManager/ViewModels/ProfilesViewModel.cs
@@ -11,6 +11,7 @@ using System.Collections.Generic;
 using System.Collections;
 using System.Windows.Threading;
 using NETworkManager.Enum;
+using NETworkManager.Models.Profile;
 
 namespace NETworkManager.ViewModels
 {
