+++ Source/NETworkManager/ViewModels/TigerVNCHostViewModel.cs
@@ -15,6 +15,7 @@ using NETworkManager.Utilities;
 using System.Windows;
 using NETworkManager.Models.TigerVNC;
 using NETworkManager.Models.EventSystem;
+using NETworkManager.Models.Profile;
 
 namespace NETworkManager.ViewModels
 {
