+++ Source/NETworkManager/ViewModels/TracerouteHostViewModel.cs
@@ -11,6 +11,7 @@ using System.Windows.Data;
 using System.Linq;
 using MahApps.Metro.Controls.Dialogs;
 using System.Windows;
+using NETworkManager.Models.Profile;
 
 namespace NETworkManager.ViewModels
 {
